# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/setup:
    post:
      operationId: completeSetup
      tags: [auth]
      summary: Run the one-time setup wizard (create store + admin + default roles)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupInput"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupResult"
        "409":
          description: Setup already completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: Account locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [auth]
      summary: Issue a new access token using the refresh cookie
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Current authenticated user with resolved permissions
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SetupInput:
      type: object
      required:
        - storeName
        - printerWidth
        - adminFullName
        - adminUsername
        - adminPassword
      properties:
        storeName:
          type: string
          minLength: 1
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        currency:
          type: string
          default: EGP
        taxRate:
          type: number
          minimum: 0
          maximum: 100
          default: 0
        logoUrl:
          type: string
        printerWidth:
          type: string
          enum: ["58mm", "80mm"]
        paperType:
          type: string
        adminFullName:
          type: string
          minLength: 1
        adminUsername:
          type: string
          minLength: 3
        adminPassword:
          type: string
          minLength: 4
    SetupResult:
      type: object
      properties:
        storeId:
          type: string
          format: uuid
        message:
          type: string
      required:
        - storeId
        - message
    LoginInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    RoleRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        nameAr:
          type: ["string", "null"]
      required:
        - id
        - name
    CurrentUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        username:
          type: string
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/RoleRef"
        permissions:
          type: array
          items:
            type: string
        storeName:
          type: string
      required:
        - id
        - storeId
        - username
        - fullName
        - role
        - permissions
        - storeName
    AuthResult:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/CurrentUser"
      required:
        - accessToken
        - user
...